Moved helper to query folder | Removed throw exception, replaced with message on page
