+++ src/Intracto/SecretSantaBundle/Resources/config/services.xml
@@ -3,9 +3,8 @@
 <container xmlns="http://symfony.com/schema/dic/services"
            xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
            xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
-
     <services>
-        <service id="intracto_secret_santa.reporting" class="Intracto\SecretSantaBundle\Helper\ReportingHelper">
+        <service id="intracto_secret_santa.reporting" class="Intracto\SecretSantaBundle\Query\ReportQuery">
             <argument type="service" id="doctrine"></argument>
         </service>
     </services>
