+++ src/Intracto/SecretSantaBundle/Resources/translations/messages.en.yml
@@ -373,3 +373,4 @@ report:
     user_amount: 'Total amount of users'
     average_user_amount: 'Average amount of users per pool'
     average_wishlist_amount: 'Average amount of completed wishlists'
+    no_data: 'There are no data available.'
