+++ src/Intracto/SecretSantaBundle/Resources/views/Report/report.html.twig
@@ -12,36 +12,42 @@
                 <select name="year" id="Year" onchange="this.form.submit();">
                     <option {% if(current_year == "all") %} selected="selected" {% endif %} value="all">All time
                     </option>
-                    {% for year in dataPool['featured_years'] %}
-                        <option value="{{ year.featured_year }}" {% if(current_year == year.featured_year) %} selected="selected" {% endif %}>{{ year.featured_year }}</option>
+                    {% for year in featured_years['featured_years'] %}
+                        <option value="{{ year }}" {% if(current_year == year) %} selected="selected" {% endif %}>{{ year }}</option>
                     {% endfor %}
                 </select>
             </div>
         </div>
     </form>
 
-    <div class="col-xs-6 .col-sm-3 report">
-        <h2>{{ 'report.pool_amount'|trans }}</h2>
+    {% if dataPool is empty %}
+        <div>
+            <p>{{ 'report.no_data'|trans }}</p>
+        </div>
+    {% else %}
+        <div class="col-xs-6 .col-sm-3 report">
+            <h2>{{ 'report.pool_amount'|trans }}</h2>
 
-        <p><strong>{{ dataPool['pools'][0]['poolCount'] }}</strong></p>
-    </div>
+            <p><strong>{{ dataPool['pools'][0]['poolCount'] }}</strong></p>
+        </div>
 
-    <div class="col-xs-6 .col-sm-3 report">
-        <h2>{{ 'report.user_amount'|trans }}</h2>
+        <div class="col-xs-6 .col-sm-3 report">
+            <h2>{{ 'report.user_amount'|trans }}</h2>
 
-        <p><strong>{{ dataPool['entries'][0]['entryCount'] }}</strong></p>
-    </div>
+            <p><strong>{{ dataPool['entries'][0]['entryCount'] }}</strong></p>
+        </div>
 
-    <div class="col-xs-6 .col-sm-3 report">
-        <h2>{{ 'report.average_user_amount'|trans }}</h2>
+        <div class="col-xs-6 .col-sm-3 report">
+            <h2>{{ 'report.average_user_amount'|trans }}</h2>
 
-        <p><strong>{{ dataPool['entry_average'] }}</strong></p>
-    </div>
+            <p><strong>{{ dataPool['entry_average'] }}</strong></p>
+        </div>
 
-    <div class="col-xs-6 .col-sm-3 report">
-        <h2>{{ 'report.average_wishlist_amount'|trans }}</h2>
+        <div class="col-xs-6 .col-sm-3 report">
+            <h2>{{ 'report.average_wishlist_amount'|trans }}</h2>
 
-        <p><strong>{{ dataPool['wishlist_average'] }}%</strong></p>
-    </div>
-    <div class="clearfix"></div>
+            <p><strong>{{ dataPool['wishlist_average'] }}%</strong></p>
+        </div>
+        <div class="clearfix"></div>
+    {% endif %}
 {% endblock %}
